+++ tests/fixtures/all/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  extends: '../../../lib/config/all.js',
+  settings: {
+    react: {
+      version: '16.0',
+    },
+  },
+};