+++ tests/fixtures/all/app/foo.js
@@ -1,0 +1,4 @@
+/* eslint-disable flowtype/require-valid-file-annotation */
+module.exports = {
+  foo: 'bar',
+};