+++ tests/fixtures/all/app/index.js
@@ -1,0 +1,7 @@
+/* eslint-disable flowtype/require-valid-file-annotation,typescript/no-var-requires */
+const {foo} = require('./foo');
+
+module.exports = {
+  foo,
+  bar: 'bar',
+};