+++ tests/fixtures/prettier-graphql/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [
+    '../../../lib/config/graphql.js',
+    '../../../lib/config/prettier.js',
+  ],
+};