+++ tests/fixtures/typescript-no-js/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [
+    '../../../lib/config/node.js',
+    '../../../lib/config/typescript.js',
+  ],
+};