+++ tests/lib/config/prettier.js
@@ -1,24 +1,57 @@
 /* eslint-env node, mocha */
-/* eslint-disable flowtype/require-valid-file-annotation */
 import {expect} from 'chai';
-import {execSync} from 'child_process';
 
 // eslint-disable-next-line shopify/strict-component-boundaries
-import {fixtureFile} from '../../utilities';
+import {configFile, execESLint, fixtureFile} from '../../utilities';
 
 describe('config', () => {
   describe('prettier', () => {
     it('validates source files using prettier', () => {
       expect(
-        () =>
-          execSync(
-            `node_modules/.bin/eslint --no-ignore ${fixtureFile('prettier')}`,
-          ),
-        // eslint-disable-next-line function-paren-newline
-      )
-        .throws()
-        .with.property('stdout')
-        .to.match(/Replace .*"bar".* with .*'bar'/);
+        execESLint(
+          `--config ${configFile('prettier')} ${fixtureFile('prettier')}`,
+        ),
+      ).to.match(/Replace .*"bar".* with .*'bar'/);
+    }).timeout(8000);
+
+    it('validates TypeScript source files using prettier', () => {
+      expect(
+        execESLint(
+          `--ext .ts --config ${fixtureFile(
+            'prettier-typescript/.eslintrc.js',
+          )} ${fixtureFile('prettier-typescript')}`,
+        ),
+      ).to.match(/Replace .*"bar".* with .*'bar'/);
+    }).timeout(8000);
+
+    it('does not generate prettier errors in graphql files', () => {
+      expect(
+        execESLint(
+          `--ext .graphql --ignore-pattern "**/prettier-graphql/build/*" --config "${fixtureFile(
+            'prettier-graphql/.eslintrc.js',
+          )}" "${fixtureFile('prettier-graphql')}"`,
+        ),
+      ).to.eq('');
+    }).timeout(8000);
+
+    it('uses .prettierrc as the source of prettier rules', () => {
+      expect(
+        execESLint(
+          `--config ${configFile('prettier')} ${fixtureFile(
+            'prettier-config',
+          )}`,
+        ),
+      ).to.match(/Delete .;./);
+    }).timeout(8000);
+
+    it('uses .prettierrc as the source of prettier rules for TypeScript', () => {
+      expect(
+        execESLint(
+          `--ext .ts --plugin typescript --config ${fixtureFile(
+            'prettier-typescript/.eslintrc.js',
+          )} ${fixtureFile('prettier-config')}`,
+        ),
+      ).to.match(/Delete .;./);
     }).timeout(8000);
   });
 });